Added a bit of testing to verify the behavior.
